+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
 
+## [0.1.0-beta] - Feb 26, 2025
+- Optimized Milstones
+- Added additional test for Milestone removal
+- Adjust namespace to be more consistent with other Nonatomic packages
+
 ## [0.0.0-beta] - Sept 28, 2024
 - First commit